2.7.0

See merge request verifit/unite!47

+++ run_all.sh
@@ -33,10 +33,20 @@ ROOTDIR="$(dirname "$(realpath "$0")")" # get the script directory
 # source shared utils
 source "$ROOTDIR/dev_tools/shared.sh"
 
+# variables used to control which log tailing terminals
+# can be killed during a failed startup
+# (i.e. their corresponding component started succesfully)
+unset KILL_TRIPLESTORE_TAIL KILL_COMPILATION_TAIL KILL_ANALYSIS_TAIL
 killTailTerminals(){
-    kill "$(ps -ef | grep "tail -f .*/logs/../logs/../logs/analysis_" | head -1 | awk '{ print $2 }')" &> /dev/null     # funny path /logs/../logs/../logs/ to avoid killing unwanted tail commands 
-    kill "$(ps -ef | grep "tail -f .*/logs/../logs/../logs/compilation_" | head -1 | awk '{ print $2 }')" &> /dev/null  # TODO use PIDs instead ($!)
-    kill "$(ps -ef | grep "tail -f .*/logs/../logs/../logs/triplestore_" | head -1 | awk '{ print $2 }')" &> /dev/null
+    if [ -n "$KILL_ANALYSIS_TAIL" ] || [ $1 -eq 1 ]; then
+        kill "$(ps -ef | grep "tail -f .*/logs/../logs/../logs/analysis_" | head -1 | awk '{ print $2 }')" &> /dev/null     # funny path /logs/../logs/../logs/ to avoid killing unwanted tail commands
+    fi
+    if [ -n "$KILL_COMPILATION_TAIL" ] || [ $1 -eq 1 ]; then
+        kill "$(ps -ef | grep "tail -f .*/logs/../logs/../logs/compilation_" | head -1 | awk '{ print $2 }')" &> /dev/null  # TODO use PIDs instead ($!)
+    fi
+    if [ -n "$KILL_TRIPLESTORE_TAIL" ] || [ $1 -eq 1 ]; then
+        kill "$(ps -ef | grep "tail -f .*/logs/../logs/../logs/triplestore_" | head -1 | awk '{ print $2 }')" &> /dev/null
+    fi
 }
 
 # catch ctrl+c and kill all subprocesses
@@ -46,7 +56,21 @@ killall() {
     echo -e "\nShutting down..."
     kill -TERM 0
     if [ -n "$ARG_TAIL" ]; then
-        killTailTerminals
+        killTailTerminals 1
+    fi
+    wait
+    echo "All done."
+
+    exit 0
+}
+
+# used by the script in case of failure
+abortAll() {
+    trap '' INT TERM     # ignore INT and TERM while shutting down
+    echo -e "\nShutting down..."
+    kill -TERM 0
+    if [ -n "$ARG_TAIL" ]; then
+        killTailTerminals 0
     fi
     wait
     echo "All done."
@@ -92,7 +116,7 @@ main () {
     # get and output version
     VERSION="$(cat "$ROOTDIR/VERSION.md" 2>/dev/null)"
     echo -e "\n########################################################"
-    echo -e "    ${BOLD}Unite${NORMAL}, ${VERSION}"
+    echo -e "    ${BOLD}Unite${NORMAL} ${VERSION}"
     echo -e "########################################################\n"
 
     # lookup config URLs
@@ -112,68 +136,114 @@ main () {
     for log in `ls | grep triplestore_.*\.log | head -n -5` ; do rm $log ; done
     for log in `ls | grep compilation_.*\.log | head -n -5` ; do rm $log ; done
     for log in `ls | grep analysis_.*\.log | head -n -5` ; do rm $log ; done
-    cd -
+    cd - &> /dev/null
 
-    # open new terminals that tail the log files and record their PIDs to kill later
+    ############################################################################################################
+    #  Triplestore
+    ############################################################################################################
+
+    # open new terminal that tails the log file and record its PID to kill later
     if [ -n "$ARG_TAIL" ]; then
         gnome-terminal --title="tail: Triplestore Log (feel free to close this window)" -- /bin/bash -c "tail -f \"$ROOTDIR/logs/../logs/../logs/triplestore_$CURTIME.log\"" # funny path /logs/../logs/../logs/ to avoid killing unwanted tail commands 
-        gnome-terminal --title="tail: Compilation Log (feel free to close this window)" -- /bin/bash -c "tail -f \"$ROOTDIR/logs/../logs/../logs/compilation_$CURTIME.log\"" # TODO use PIDs instead ($!)
-        gnome-terminal --title="tail: Analysis Log (feel free to close this window)" -- /bin/bash -c "tail -f \"$ROOTDIR/logs/../logs/../logs/analysis_$CURTIME.log\""
     fi
 
     # start the triplestore
-    echo "Starting the Triplestore"
+    echo -n "Starting the Triplestore"
     "$ROOTDIR/sparql_triplestore/run.sh" &> "$ROOTDIR/logs/triplestore_$CURTIME.log" &
     PROCESS_PID=$!
+    echo " ($PROCESS_PID)"
     echo -e "Waiting for the Triplestore to finish startup"
     waitForUrlOnline "$triplestore_url" "$PROCESS_PID" "$SLEEP" 0
     ret="$?"
     if [ "$ret" -eq 0 ]; then
         echo -e "Triplestore ${GREEN}running${NC} (1/3)\n"
+        KILL_TRIPLESTORE_TAIL=true
     else
-        echo -e "Triplestore ${RED}failed${NC} to start!\n"
-        echo -e "Try checking the log $ROOTDIR/logs/triplestore_$CURTIME.log\nor run again with '-t' to see the logs during startup.\n"
-        killall # exit
+        echo -e "Triplestore ${RED}failed${NC} to start!"
+        if [ -n "$ARG_TAIL" ]; then
+            echo -e "See the \"tail: Triplestore Log\" terminal window for the startup log."
+            echo -e "Or try checking the logs: \"$ROOTDIR/logs/triplestore_$CURTIME.log\""
+        else
+            echo -e "Try checking the logs: \"$ROOTDIR/logs/triplestore_$CURTIME.log\""
+            echo -e "Or run again with '-t' to see the logs during startup."
+        fi
+        abortAll # exit
+    fi
+
+
+    ############################################################################################################
+    #  Compilation
+    ############################################################################################################
+
+    # open new terminal that tails the log file and record its PID to kill later
+    if [ -n "$ARG_TAIL" ]; then
+        gnome-terminal --title="tail: Compilation Log (feel free to close this window)" -- /bin/bash -c "tail -f \"$ROOTDIR/logs/../logs/../logs/compilation_$CURTIME.log\"" # TODO use PIDs instead ($!)
     fi
 
     # start the compilation adapter
-    echo "Starting the Compilation adapter"
+    echo -n "Starting the Compilation adapter"
     cd "$ROOTDIR/compilation"
     mvn jetty:run-exploded &> "$ROOTDIR/logs/compilation_$CURTIME.log" &
     PROCESS_PID=$!
-    echo $PROCESS_PID
+    echo " ($PROCESS_PID)"
     cd - > /dev/null
     echo -e "Waiting for the Compilation adapter to finish startup"
     waitForUrlOnline "$compilation_url" "$PROCESS_PID" "$SLEEP" 0
     ret="$?"
     if [ "$ret" -eq 0 ]; then
         echo -e "Compilation adapter ${GREEN}running${NC} (2/3)\n"
+        KILL_COMPILATION_TAIL=true
     else
-        echo -e "Compilation adapter ${RED}failed${NC} to start!\n"
-        echo -e "Try checking the log $ROOTDIR/logs/compilation_$CURTIME.log\nor run again with '-t' to see the logs during startup.\n"
-        killall # exit
+        echo -e "Compilation adapter ${RED}failed${NC} to start!"
+        if [ -n "$ARG_TAIL" ]; then
+            echo -e "See the \"tail: Compilation Log\" terminal window for the startup log."
+            echo -e "Or try checking the logs: \"$ROOTDIR/logs/compilation_$CURTIME.log\""
+        else
+            echo -e "Try checking the logs: \"$ROOTDIR/logs/compilation_$CURTIME.log\""
+            echo -e "Or run again with '-t' to see the logs during startup."
+        fi
+        abortAll # exit
+    fi
+
+
+    ############################################################################################################
+    #  Analysis
+    ############################################################################################################
+
+    # open new terminal that tails the log file and record its PID to kill later
+    if [ -n "$ARG_TAIL" ]; then
+        gnome-terminal --title="tail: Analysis Log (feel free to close this window)" -- /bin/bash -c "tail -f \"$ROOTDIR/logs/../logs/../logs/analysis_$CURTIME.log\""
     fi
 
     # start the analysis adapter
-    echo "Starting the Analysis adapter"
+    echo -n "Starting the Analysis adapter"
     cd "$ROOTDIR/analysis"
     mvn jetty:run-exploded &> "$ROOTDIR/logs/analysis_$CURTIME.log" &
     PROCESS_PID=$!
+    echo " ($PROCESS_PID)"
     cd - > /dev/null
     echo -e "Waiting for the Analysis adapter to finish startup"
     waitForUrlOnline "$analysis_url" "$PROCESS_PID" "$SLEEP" 0
     ret="$?"
     if [ "$ret" -eq 0 ]; then
         echo -e "Analysis adapter ${GREEN}running${NC} (3/3)\n"
+        KILL_ANALYSIS_TAIL=true
     else
-        echo -e "Analysis adapter ${RED}failed${NC} to start!\n"
-        echo -e "Try checking the log $ROOTDIR/logs/analysis_$CURTIME.log\nor run again with '-t' to see the logs during startup.\n"
-        killall # exit
+        echo -e "Analysis adapter ${RED}failed${NC} to start!"
+        if [ -n "$ARG_TAIL" ]; then
+            echo -e "See the \"tail: Analysis Log\" terminal window for the startup log."
+            echo -e "Or try checking the logs: \"$ROOTDIR/logs/analysis_$CURTIME.log\""
+        else
+            echo -e "Try checking the logs: \"$ROOTDIR/logs/analysis_$CURTIME.log\""
+            echo -e "Or run again with '-t' to see the logs during startup."
+        fi
+        abortAll # exit
     fi
 
     echo -e "${GREEN}Ready to go!${NC}"
     echo "Use ctrl+c to exit..."
-    cat # wait forever
+    wait # wait for jobs to finish (forever)
+    killall # make sure all terminals get closed and the exit message is printed
 }
 
 main "$@"
